+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## CHANGELOG
 
-1. post award span wont be rendered if there are no awards.
+1. all sub names will always be lowercased now.
+2. added PostLoader component.
+3. added `backspace` to go back in history.
+4. tried to add AbortController.
+5. fixed the vert split layout in focus mode.